pyproject.toml
[project]
name = "currency_rate_extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "paramiko>=3.5.0",
    "matplotlib>=3.9.2",
    "plotly>=5.24.1",
    "dash>=2.18.2",
    "nbformat>=5.10.4",
    "statsmodels>=0.14.4",
    "kaleido==0.2.1",
    "currency_rate_extractor",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true
package = true

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest",
]
